♻️ refactor : SupabaseAuth 파일 내용을 훅으로 만들고 기존 파일 삭제
